requirements.txt
pynacl
youtube-dl
discord
python-dotenv
pymongo
dnspython
asyncio